client/requirements.txt
pyyaml
PyQt5
chardet
pyyaml
requests
pytest
pytest-cov
fabric3
sqlalchemy
pycryptodome